Moved shared pedestrian helper functions to utility files. Fixed bounding box orientation issues for yolo and pointpillars files. Updated README with visualization instructions
